dfa: correct pre-conditions
I don't need set of all of the keys of delta's vals to be equal to
Σ; I need *each* of the set of keys of delta's vals to be equal to
Σ.

Also, the final K in K → Σ → K does not need to completely covered; it's
perfectly acceptable to have p,q ∈ K such that ∀σ ∈ Σ : δ(p,σ) ≠ q. In
other words, it's acceptable to have q unreachable from p! But all
states must have a transition function, and each state's function must
accept every symbol in the alphabet.