dfa: optimize naïve equivalent-states
The brunt of the optimization relies on the symmetric nature of the ≡
relation. If you draw a matrix with a row and column for each p ∈ K,
then you only have to check one triangular (lower or upper) subset of
the matrix (where a cell represents a pair (p,q) ∈ K^2)!

So, we do the functional equivalent of:

  Let m = {}
  For all p ∈ K:
    For (k → v) ∈ m:
      If p ≡ k:
        Add p to v
        Break
    Else (p does not belong to the equivalence class of anything in m):
      Add (p → {p}) to m
  Return the set of values of m

This should be O(|Σ||K|^2), the theoretical best according to Lewis and
Papadimitriou.